feat: change stock field type from int to float in create and update ration stock DTOs

+++ Antonieta.Backend/dtos/create_ration_stock_dto.py
@@ -7,4 +7,4 @@ class create_ration_stock_dto(BaseModel):
     name: str
     description: str
     unit: str
-    stock: int
+    stock: float
